[UPD] - Organización de vistas
Organicé mejor las vistas en las carpetas que ya habías sido creadas con anterioridad, además de crear el controlador para acceder a las vistas, ahí creé funciones para las mismas vistas y también las rutas en web para acceder a esas funciones.
