--- resources/views/Agregar-Curso.blade.php
@@ -1,110 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Agregar Curso</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-    <style>
-        body {
-            background-color: #ffffff;
-            font-family: Arial, sans-serif;
-        }
-
-        .container-div {
-            height: 100vh;
-            background-color: #ffffff;
-            border-radius: 15px;
-            box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-            padding: 20px;
-        }
-
-        .header-section {
-            background-color: #e1b8b8;
-            padding: 30px 0;
-            border-radius: 10px;
-            box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-            text-align: center;
-            margin-bottom: 20px;
-        }
-
-        .header-section h1 {
-            color: #ffffff;
-        }
-
-        .form-floating input,
-        .form-floating textarea {
-            border: 2px solid #000000;
-            border-radius: 10px;
-        }
-
-        .form-floating label {
-            color: #000000;
-        }
-
-        .btn-dark {
-            background-color: #e1b8b8;
-            border-color: #e1b8b8;
-        }
-
-        .btn-dark:hover {
-            background-color: #ff6699;
-            border-color: #ff3366;
-        }
-    </style>
-</head>
-<body>
-    <div class="container container-div">
-        <div class="header-section">
-            <h1>Cursos</h1>
-        </div>
-        
-        <div class="row">
-            <div class="col-md-12">
-                <div class="form-floating mb-3">
-                    <input type="text" class="form-control" id="nombre" name="nombre" placeholder="-- Nombre --">
-                    <label for="nombre">Nombre</label>
-                </div>
-            </div>
-            <div class="col-md-12">
-                <div class="form-floating mb-3">
-                    <input type="number" class="form-control" id="cupoLimite" name="cupoLimite" placeholder="-- Cupo Limite --">
-                    <label for="cupoLimite">Cupo Limite</label>
-                </div>
-            </div>
-            <div class="col-md-12">
-                <div class="form-floating mb-3">
-                    <input type="date" class="form-control" id="fechaInicio" name="fechaInicio">
-                    <label for="fechaInicio">Fecha de inicio</label>
-                </div>
-            </div>
-            <div class="col-md-12">
-                <div class="form-floating mb-3">
-                    <input type="time" class="form-control" id="horaInicio" name="horaInicio" placeholder="-- Hora de inicio del curso --">
-                    <label for="horaInicio">Hora de inicio</label>
-                </div>
-            </div>
-            <div class="col-md-12">
-                <div class="form-floating mb-3">
-                    <input type="number" class="form-control" id="precio" name="precio" placeholder="-- Precio --">
-                    <label for="precio">Precio</label>
-                </div>
-            </div>
-            <div class="col-md-12">
-                <div class="form-floating mb-3">
-                    <input type="text" class="form-control" id="descripcion" name="descripcion" placeholder="-- Descripcion --">
-                    <label for="descripcion">Descripción</label>
-                </div>
-            </div>
-            <div class="col-md-12 text-center">
-                <button class="btn btn-dark w-100">Agregar curso</button>
-            </div>
-        </div>
-    </div>
-    
-
-    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.5.4/dist/umd/popper.min.js"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
-</body>
-</html>
