--- resources/views/Boos/Home-Boss.blade.php
@@ -1,3 +1,0 @@
-<div>
-    <!-- Knowing is not enough; we must apply. Being willing is not enough; we must do. - Leonardo da Vinci -->
-</div>
