+++ resources/views/Boss/Ver-Inventario.blade.php
