--- resources/views/Employe/Home-Empleado.blade.php
@@ -1,3 +1,0 @@
-<div>
-    <!-- Act only according to that maxim whereby you can, at the same time, will that it should become a universal law. - Immanuel Kant -->
-</div>
